refactor: prep balance gql query
