+++ cala-ledger/src/balance/error.rs
@@ -1,6 +1,7 @@
 use thiserror::Error;
 
 use crate::outbox::error::OutboxError;
+use cala_types::primitives::*;
 
 #[derive(Error, Debug)]
 pub enum BalanceError {
@@ -12,4 +13,6 @@ pub enum BalanceError {
     EntityError(#[from] crate::entity::EntityError),
     #[error("BalanceError - OptimisticLockingError")]
     OptimisticLockingError,
+    #[error("BalanceError - NotFound: there is no balance recorded for journal {0}, account {1}, currency {2}")]
+    NotFound(JournalId, AccountId, Currency),
 }
