+++ cala-server/Cargo.toml
@@ -25,6 +25,7 @@ axum = { workspace = true }
 axum-extra = { workspace = true }
 base64 = { workspace = true }
 chrono = { workspace = true }
+rust_decimal = { workspace = true }
 derive_builder = { workspace = true }
 futures = { workspace = true }
 thiserror = { workspace = true }
